refactor: revamp CLI commands
