--- lib/mihari/cli/validator.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-require "mihari/commands/validator"
-
-module Mihari
-  module CLI
-    class Validator < Base
-      include Mihari::Commands::Validator
-    end
-  end
-end
